Move GossipLogger classes to a dedicated package
